chore: move whoami logic to cluster as it is used only there
